refactor: move all types to shared layer
