+++ src/shared/model/types.ts
