+++ src/widgets/PostList/model/hooks/usePostFilter.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback, useMemo } from "react";
-import type { Post } from "../../../../entities/post/model/types";
+import type { Post } from "../../../../shared/model/types";
 import { filterByLength } from "../../../../features/PostLengthFilter/lib/filterByLength";
 
 type UsePostFilterResult = {
